+++ apps/api/tests/test_sendgrid_handler.py
@@ -1,16 +1,11 @@
-from unittest.mock import AsyncMock, call, patch
+from unittest.mock import AsyncMock, patch
 
 import pytest
 from aiosendgrid import AsyncSendGridClient
 from httpx import HTTPStatusError, Request, Response
 
-from models.ApplicationData import Decision
 from services import sendgrid_handler
-from services.sendgrid_handler import (
-    ApplicationUpdatePersonalization,
-    ConfirmationPersonalization,
-    Template,
-)
+from services.sendgrid_handler import ConfirmationPersonalization, Template
 
 SAMPLE_SENDER = ("noreply@irvinehacks.com", "No Reply IrvineHacks")
 SAMPLE_RECIPIENTS: list[ConfirmationPersonalization] = [
@@ -142,32 +137,3 @@ async def test_sendgrid_error_causes_runtime_error(mock_AsyncClient: AsyncMock) 
             SAMPLE_RECIPIENTS,
             True,
         )
-
-
-@patch("services.sendgrid_handler.send_email")
-async def test_send_decision_email(mock_sendgrid_handler_send_email: AsyncMock) -> None:
-    applicants = {
-        ("test1", "test1@uci.edu"): Decision.ACCEPTED,
-        ("test2", "test2@uci.edu"): Decision.REJECTED,
-        ("test3", "test3@uci.edu"): Decision.WAITLISTED,
-    }
-
-    accepted = [
-        ApplicationUpdatePersonalization(first_name="test1", email="test1@uci.edu")
-    ]
-    rejected = [
-        ApplicationUpdatePersonalization(first_name="test2", email="test2@uci.edu")
-    ]
-    waitlisted = [
-        ApplicationUpdatePersonalization(first_name="test3", email="test3@uci.edu")
-    ]
-
-    await sendgrid_handler.send_decision_email(SAMPLE_SENDER, applicants)
-
-    mock_sendgrid_handler_send_email.assert_has_calls(
-        [
-            call(Template.ACCEPTED_EMAIL, SAMPLE_SENDER, accepted, True),
-            call(Template.REJECTED_EMAIL, SAMPLE_SENDER, rejected, True),
-            call(Template.WAITLISTED_EMAIL, SAMPLE_SENDER, waitlisted, True),
-        ]
-    )
